+++ docs/xma.optimizers.sgd.param_group.rst
@@ -1,0 +1,7 @@
+xma.optimizers.sgd.param\_group
+===============================
+
+.. automodule:: xma.optimizers.sgd.param_group
+   :members:
+   :show-inheritance:
+   :undoc-members:
